% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{element}
\alias{element}
\title{Element}
\format{
A data frame with four variables:
\describe{
\item{\code{elementId}}{A unique string identifying the element.}
\item{\code{stratum}}{Name of the stratum that the element belongs to.}
\item{\code{latitude}}{Latitude in decimal degrees.}
\item{\code{longitude}}{Longitude in decimal degrees.}
}
}
\usage{
element
}
\description{
Example of element dataset that is used by
\code{\link{buffered_random_sampling}}.
}
\keyword{datasets}
