% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stratum}
\alias{stratum}
\title{Stratum}
\format{
A data frame with three variables:
\describe{
\item{\code{stratum}}{A unique string identifying the stratum.}
\item{\code{area}}{Area of stratum in \eqn{m^2}.}
\item{\code{n_stations}}{Required number of stations in stratum.}
}
}
\usage{
stratum
}
\description{
Example of statum dataset that is used by
\code{\link{buffered_random_sampling}}.
}
\keyword{datasets}
