% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buffered_random_sampling.R
\name{buffered_random_sampling}
\alias{buffered_random_sampling}
\title{Buffered Random Sampling}
\usage{
buffered_random_sampling(element, stratum)
}
\arguments{
\item{element}{Data set with elements. See \code{\link{element}}.}

\item{stratum}{Data set with stratums. See \code{\link{stratum}}.}
}
\value{
Data frame listing which stratum each element is associated with.
}
\description{
Run the buffered random sampling algorithm.
}
\details{
The function requires two data sets and returns a data frame with the result
of the buffered random sampling.
}
