% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buffered_random_sampling.R
\name{buffered_random_sampling}
\alias{buffered_random_sampling}
\title{Buffered Random Sampling}
\usage{
buffered_random_sampling(
  element,
  stratum,
  visualise = FALSE,
  verbose = FALSE,
  pause = FALSE,
  detail = 2
)
}
\arguments{
\item{element}{Data frame with elements. See \code{\link{element}}.}

\item{stratum}{Data frame with stratums. See \code{\link{stratum}}.}

\item{visualise}{If TRUE a plot of the current stage of the algorithm is
shown.}

\item{verbose}{Print status while the algorithm is running.}

\item{pause}{Stop the algorithm to inspect progress. Makes most sense when
used together with \code{visualise} or \code{verbose}.}

\item{detail}{Control how often \code{visualise}, \code{verbose}, and
\code{pause} are shown, printed, or executed. Possible values 1, 2, 3, and 4,
corresponding to finish, after each stratum, each time the algorithm is
reset, or after every element.}
}
\value{
Data frame listing which stratum each element is associated with.
}
\description{
Run the buffered random sampling algorithm.
}
\details{
The function requires two data sets and returns a data frame with the result
of the buffered random sampling.
}
