% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buffered_random_sampling.R
\name{buffered_random_sampling}
\alias{buffered_random_sampling}
\title{Buffered Random Sampling}
\usage{
buffered_random_sampling(
  element,
  stratum,
  preselect_element = NULL,
  visualise = FALSE,
  verbose = FALSE,
  pause = FALSE,
  detail = 2
)
}
\arguments{
\item{element}{\code{data.frame} See \code{\link{element}}.}

\item{stratum}{\code{date.frame} See \code{\link{stratum}}.}

\item{preselect_element}{\code{data.frame} Vector of elementIds to mark as
selected before the algorithm starts running.}

\item{visualise}{\code{logical} If TRUE a plot of the current stage of the
algorithm is shown.}

\item{verbose}{\code{logical} Print status while the algorithm is running.}

\item{pause}{\code{logical} Stop the algorithm to inspect progress. Makes
most sense when used together with \code{visualise} or \code{verbose}.}

\item{detail}{\code{numeric}. Control how often \code{visualise},
\code{verbose}, and \code{pause} are shown, printed, or executed. Possible
values 1, 2, 3, and 4, corresponding to finish, after each stratum, each time
the algorithm is reset, or after every element.}
}
\value{
Data frame listing which stratum each element is associated with.
}
\description{
Run the buffered random sampling algorithm.
}
\details{
The function requires two data sets and returns a data frame with the result
of the buffered random sampling.
}
